Quick work on D3D12
Implementation of command buffer

+++ Source/Rendering/ICrCommandQueue.h
@@ -20,7 +20,7 @@ namespace CrCommandQueueType
 	// | |__________| |
 	// |______________|
 
-	enum T : uint8_t
+	enum T : uint32_t
 	{
 		Graphics,
 		Compute,
